Fix Claims.Validate
The `within()` function is wrong since it is supposed to check
`max-delta < val < max+delta` but not `val == max`. It also does not
work for the `nbf` check since that needs to verify that `nbf-leeway <
now`

Also, if I read RFC 7519 correctly then the token should not be used
after `exp` and not before `nbf`. So `Validate` should check for `nbf <
now <= exp` or more generally `nbf-x < now <= exp+y`