jwt_test: Fix test assertions to use || instead of &&
Two of the test cases in jwt_test.go perform several checks against the
claimset but require that all of them be true (&&) in order to trigger
the assert. More than likely, this is not the desired behavior and the
error should actually be triggered if any (||) of the conditions is
true.

This change modifies TestJWTValidator and TestBasicJWT to use || instead
of && to chain the conditionals in their assertion.